test: carousel 스토리북 변경
